+++ crates/reqwest_client/Cargo.toml
@@ -24,6 +24,7 @@ anyhow.workspace = true
 bytes.workspace = true
 futures.workspace = true
 http_client.workspace = true
+http_client_tls.workspace = true
 serde.workspace = true
 smol.workspace = true
 log.workspace = true
